Auth repository in progress
