+++ app/src/main/AndroidManifest.xml
@@ -13,7 +13,7 @@
         android:theme="@style/Theme.WhatsAppCompose"
         tools:targetApi="31">
         <activity
-            android:name=".presentation.MainActivity"
+            android:name=".core.presentation.MainActivity"
             android:exported="true"
             android:label="@string/app_name"
             android:theme="@style/Theme.WhatsAppCompose">
