+++ app/src/main/java/com/example/whatsappcompose/auth/domain/AuthError.kt
@@ -1,0 +1,7 @@
+package com.example.whatsappcompose.auth.domain
+
+import com.example.whatsappcompose.core.domain.Error
+
+sealed interface AuthError: Error {
+    data object LoginError: AuthError
+}
