+++ app/src/main/java/com/example/whatsappcompose/auth/domain/repository/AuthRepository.kt
@@ -1,0 +1,8 @@
+package com.example.whatsappcompose.auth.domain.repository
+
+import com.google.firebase.auth.AuthResult
+import kotlinx.coroutines.flow.Flow
+
+interface AuthRepository {
+    fun login(email: String, password: String): Flow<AuthResult>
+}
