+++ app/src/main/java/com/example/whatsappcompose/auth/presentation/AuthViewModel.kt
@@ -1,0 +1,7 @@
+package com.example.whatsappcompose.auth.presentation
+
+import androidx.lifecycle.ViewModel
+
+class AuthViewModel: ViewModel() {
+
+}
