+++ app/src/main/java/com/example/whatsappcompose/auth/presentation/SignUpScreen.kt
@@ -1,4 +1,4 @@
-package com.example.whatsappcompose.presentation
+package com.example.whatsappcompose.auth.presentation
 
 import androidx.compose.foundation.Image
 import androidx.compose.foundation.layout.Arrangement
@@ -11,47 +11,52 @@ import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.layout.size
 import androidx.compose.foundation.text.KeyboardOptions
 import androidx.compose.material.icons.Icons
+import androidx.compose.material.icons.automirrored.rounded.ArrowBack
 import androidx.compose.material.icons.rounded.Email
 import androidx.compose.material.icons.rounded.Lock
+import androidx.compose.material.icons.rounded.Person
 import androidx.compose.material3.Button
+import androidx.compose.material3.ExperimentalMaterial3Api
 import androidx.compose.material3.Icon
 import androidx.compose.material3.IconButton
 import androidx.compose.material3.OutlinedTextField
 import androidx.compose.material3.Scaffold
 import androidx.compose.material3.Text
-import androidx.compose.material3.TextButton
+import androidx.compose.material3.TopAppBar
+import androidx.compose.material3.TopAppBarDefaults
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.getValue
 import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.remember
-import androidx.compose.runtime.saveable.rememberSaveable
 import androidx.compose.runtime.setValue
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.res.stringResource
-import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.text.input.KeyboardType
 import androidx.compose.ui.text.input.PasswordVisualTransformation
 import androidx.compose.ui.text.input.VisualTransformation
 import androidx.compose.ui.unit.dp
-import androidx.compose.ui.unit.sp
 import com.example.whatsappcompose.R
-import com.example.whatsappcompose.navigation.Screens
 import com.example.whatsappcompose.ui.theme.DarkGreen
-import com.example.whatsappcompose.util.UiEvent
+import com.example.whatsappcompose.core.util.UiEvent
 
+@OptIn(ExperimentalMaterial3Api::class)
 @Composable
-fun LoginScreen(
-    onEvent: (UiEvent.Navigate) -> Unit
+fun SignUpScreen(
+    popBackStack: (UiEvent.PopBackStack) -> Unit
 ) {
+    val scrollBehavior = TopAppBarDefaults.pinnedScrollBehavior()
+    var name by remember {
+        mutableStateOf("")
+    }
     var email by remember {
         mutableStateOf("")
     }
     var password by remember {
         mutableStateOf("")
     }
-    var isPasswordHidden by rememberSaveable {
+    var isPasswordHidden by remember {
         mutableStateOf(true)
     }
     val visualTransformation =
@@ -61,7 +66,28 @@ fun LoginScreen(
 
     Scaffold(
         modifier = Modifier
-            .fillMaxSize()
+            .fillMaxSize(),
+        topBar = {
+            TopAppBar(
+                title = {
+                    Text(text = stringResource(id = R.string.signup_toolbar_title))
+                },
+                navigationIcon = {
+                    IconButton(
+                        onClick = {
+                            popBackStack(UiEvent.PopBackStack)
+                        }
+                    ) {
+                        Icon(
+                            imageVector = Icons.AutoMirrored.Rounded.ArrowBack,
+                            tint = DarkGreen,
+                            contentDescription = ""
+                        )
+                    }
+                },
+                scrollBehavior = scrollBehavior
+            )
+        }
     ) { innerPadding ->
         Column(
             modifier = Modifier
@@ -78,6 +104,29 @@ fun LoginScreen(
                 contentDescription = ""
             )
             Spacer(modifier = Modifier.height(16.dp))
+            OutlinedTextField(
+                modifier = Modifier
+                    .fillMaxWidth(),
+                value = name,
+                onValueChange = {
+                    name = it
+                },
+                label = {
+                    Text(text = stringResource(id = R.string.name_text_input))
+                },
+                leadingIcon = {
+                    Icon(
+                        imageVector = Icons.Rounded.Person,
+                        tint = DarkGreen,
+                        contentDescription = ""
+                    )
+                },
+                singleLine = true,
+                keyboardOptions = KeyboardOptions(
+                    keyboardType = KeyboardType.Text
+                )
+            )
+            Spacer(modifier = Modifier.height(8.dp))
             OutlinedTextField(
                 modifier = Modifier
                     .fillMaxWidth(),
@@ -130,7 +179,6 @@ fun LoginScreen(
                     ) {
                         Icon(
                             painter = painterResource(id = isVisibly),
-                            tint = DarkGreen,
                             contentDescription = ""
                         )
                     }
@@ -138,47 +186,14 @@ fun LoginScreen(
                 visualTransformation = visualTransformation
             )
             Spacer(modifier = Modifier.height(8.dp))
-            TextButton(
-                onClick = {
-                    onEvent(UiEvent.Navigate(Screens.ForgotPasswordScreen))
-                }
-            ) {
-                Text(text = stringResource(id = R.string.forgot_password_button))
-            }
-        }
-        Column(
-            modifier = Modifier
-                .fillMaxSize()
-                .padding(16.dp),
-            horizontalAlignment = Alignment.CenterHorizontally,
-            verticalArrangement = Arrangement.Bottom
-        ) {
             Button(
                 modifier = Modifier
                     .fillMaxWidth(),
                 onClick = {
-
+                    TODO()
                 }
             ) {
-                Text(
-                    text = stringResource(id = R.string.login_button),
-                    fontWeight = FontWeight.Bold,
-                    fontSize = 16.sp
-                )
-            }
-            Spacer(modifier = Modifier.height(8.dp))
-            Button(
-                modifier = Modifier
-                    .fillMaxWidth(),
-                onClick = {
-                    onEvent(UiEvent.Navigate(Screens.SignUpScreen))
-                }
-            ) {
-                Text(
-                    text = stringResource(id = R.string.create_account_button),
-                    fontWeight = FontWeight.Bold,
-                    fontSize = 16.sp
-                )
+                Text(text = stringResource(id = R.string.create_account_button))
             }
         }
     }
