+++ app/src/main/java/com/example/whatsappcompose/core/domain/Error.kt
@@ -1,0 +1,3 @@
+package com.example.whatsappcompose.core.domain
+
+interface Error
