+++ app/src/main/java/com/example/whatsappcompose/core/domain/Result.kt
@@ -1,0 +1,8 @@
+package com.example.whatsappcompose.core.domain
+
+typealias RootError = Error
+
+sealed interface Result<out D, out E: RootError> {
+    data class Success<out D, out E: RootError>(val data: D): Result<D, E>
+    data class Error<out D, out E: RootError>(val error: E): Result<D, E>
+}
