+++ app/src/main/java/com/example/whatsappcompose/core/navigation/Screens.kt
@@ -1,4 +1,4 @@
-package com.example.whatsappcompose.navigation
+package com.example.whatsappcompose.core.navigation
 
 import kotlinx.serialization.Serializable
 
@@ -6,9 +6,13 @@ sealed interface Screens {
     @Serializable
     data object Auth: Screens
     @Serializable
+    data object Main: Screens
+    @Serializable
     data object LoginScreen: Screens
     @Serializable
     data object SignUpScreen: Screens
     @Serializable
     data object ForgotPasswordScreen: Screens
+    @Serializable
+    data object MainScreen: Screens
 }
